tools/photo-poster/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pillow-heif
exifread
jinja2
